pyproject.toml
[project]
name = "smashgg-exporter"
version = "0.1.0"
description = ""
authors = [
    {name = "mli42",email = "mli@student.42.fr"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv (>=1.2.1,<2.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "asyncio (>=4.0.0,<5.0.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "alembic (>=1.17.2,<2.0.0)",
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
